fix: avoid retaining gsplat work buffer shaders
